+++ sora-android-sdk/src/main/kotlin/jp/shiguredo/sora/sdk/channel/SoraMediaChannel.kt
@@ -287,6 +287,9 @@ class SoraMediaChannel(
         SoraLogger.d(TAG, "connect: mediaOption.videoCapturer          = ${mediaOption.videoCapturer}")
         SoraLogger.d(TAG, "connect: mediaOption.spotlight              = ${mediaOption.spotlight}")
         SoraLogger.d(TAG, "connect: mediaOption.sdpSemantics           = ${mediaOption.sdpSemantics}")
+        if (mediaOption.planB()) {
+            SoraLogger.w(TAG, "Plan-B SDP semantics has no longer been supported. Unified plan should be used.")
+        }
 
         if (closing) {
             return