+++ CHANGES.md
@@ -16,6 +16,7 @@
 
 - libwebrtc を 71.16.0 に上げた
 - SDP semantics のデフォルトを Unified Plan に変更した
+- Kotlin を 1.3.20 に上げた
 
 ## 1.7.1
 
@@ -24,7 +25,7 @@
 - dokka を 0.9.17 に上げた
   - 不要な generated クラスの HTML が出力されなくなった
   - sora-android-sdk-doc の api doc はすでに 0.9.17 生成版で更新済み
-- Kotoin を 1.2.71 に上げた
+- Kotlin を 1.2.71 に上げた
 - `com.google.code.gson:gson` を 2.8.5 に上げた
 - `com.squareup.okhttp3:okhttp` を 3.11.0 に上げた
 - `io.reactivex.rxjava2:rxandroid` を 2.1.0 に上げた